rename 'ldk.ts' --> 'signer.ts'
